# Use docker stop instead of sigkill in fv tests