Not sure this is exactly the right fix, but wanted to submit before I go on holiday.

I'm finding that the k8s-apiserver container doesn't die in response to an os.Interrupt. My test runs end up panicking during teardown as a result.

This PR fixes my panics by issuing a stop signal using docker directly, which actually stops the k8s-apiserver pod.

related PR: https://github.com/projectcalico/k8s-policy/pull/139