@neiljerram 2017-09-18
@ozdanborne I'm inclined to be conservative and not to make this change unless we can be sure that it's needed.  In particular, the felix PR at #1551 adds a k8s-apiserver container, _and_ uses the existing Stop() implementation to stop it - and in the felix FV that seems to work reliably every time.