@fasaxc 2017-09-19
@neiljerram We use `tini` in felix/typha for that reason.  I believe newer versions of docker even have some support for tini built in, so you can run `docker run --tini` and it'll put in a `tini` shim around the process that you run.