@neiljerram 2017-09-19
@fasaxc Right, thanks.

@ozdanborne So if the container that you're using has tini, or a similar init shim, SIGINT will work; and interrupting a process that the test code has in hand feels more elegant to me than `docker stop`.