@ozdanborne 2017-09-19
Thanks for the info guys. I looked into it some more and noticed that it was etcd that was failing - not k8s-apiserver. I fixed it by pulling a more recent `etcd:latest` image. 

I've updated as per your comments. Hoping we can merge these logging improvements. 

For reference, this is what they look like now: `time="2017-09-19T13:45:08-07:00" level=info msg="About to run container" container=&{calico-policy-controller-3750-3-   <nil> {0 0} <nil>} `

I can see how logging the entire struct is more useful once its actually running: `time="2017-09-19T13:45:08-07:00" level=info msg="Container now running" container=&{calico-policy-control
ler-3750-3- 172.17.0.7 d001318c60f9 0xc420432420 {0 0} map[]} `