@ozdanborne 2017-09-20
Scratch that, I'm still seeing intermittent issues killing apiserver and etcd when using the sigint approach. 

I strongly recommend we switch to the `docker stop` approach. It has builtin support where it waits `-t` seconds for the container to respond to the signal before issuing a hard kill.