@ozdanborne 2017-09-20
How about this compromise - switch to `SIGTERM` instead, which is the sig that `docker stop` sends? (https://docs.docker.com/engine/reference/commandline/stop/)