@fasaxc 2017-10-24
I think the key thing to note here is that we're sending the interrupt to the `docker run` CLI command.  That's like sending a Ctrl-C to that command, which should tell it to stop the container that it's running.

If that's not working in all cases, maybe we should switch to `docker stop` or do belt-and-braces.  

In any case, the idea behind using `docker run` is that it helps to make sure the container gets cleaned up if the test is killed.  I guess we could name our containers according to the pid of the test process, allowing us to reap it if the process no longer exists when the next test starts, or something like that.  

As discussed, the key things we were trying to maintain in this framework are:

- parallel execution by unique naming
- avoiding test pollution by tearing down containers between tests
- automatic cleanup if test process dies