@neiljerram 2017-10-25
@ozdanborne As we seem to be struggling to agree about the docker stop vs SIGINT, how about putting the logging changes into a separate PR, so we can at least merge those?