@djlwilder 2017-10-25
Before this issue is closed I wanted to add a comment.  I am seeing the same issue on ppc64le ( etcd and felix containers not stopping in teardown).  Here are a few things I found:

-The containers do eventual stop,  (they are no longer running after the test completes). Therefore I don't think the method of stopping them is the problem.

- Running top during the test shows dockerd at times consuming up to 200% of a cpu.

- My working theory is that this is a performance issue,  the Ginkgo runs parallel test (-p) 15 in my case.  I don't see the issue on a smaller x86 box.  Reducing the number of parallel tests helps a little.
  
- WaitNotRunning() runs  "docker ps" in a tight loop,  I added a 1 second sleep in the loop (simmer to in WaitUntilRunning() ),  dockerd peek cpu usage dropped to roughly 50%.  However, the problem still occurred.

- I increased the time to wait for a container to stop from 10 seconds to 120 seconds and increased  Ginkgo's slowSpecThreshold from 40 seconds to 120 seconds.  The tests runs ok in this configuration.  

Note:  I am only running a subset of the full fv tests at this point.